pick random on a shuffled array for the viewer route
Instead of picking a new random element and we first shuffle the array
and exhaust it. This prevent duplicate webm.
We could try to ensure the array is specific to each session, but we
would have to store cookies data. Another option would be to store the
array client side, but this could leak the list of webms.